Refactoring PDH class
